requirements.txt
django
django-jquery
django-vanilla-views
django-countries